CPU_SET時のstatic_castを削除
git-svn-id: http://10.144.169.20/repos/um/branches/l7vsd-3.x-shamshel-refine@7970 1ed66053-1c2d-0410-8867-f7571e6e31d3

+++ src/session_thread_control.cpp
@@ -43,7 +43,7 @@ void	session_thread_control::upstream_run(){
 	if( 0 < num_of_core_uses ){
 		CPU_ZERO( &mask );
 		for( int i = 0; i < num_of_core_uses; ++i ){
-			CPU_SET( static_cast<int>( i ), &mask );
+			CPU_SET( i, &mask );
 		}
 		sched_setaffinity( 0, sizeof( cpu_set_t ), &mask );
 	}
@@ -99,7 +99,7 @@ void	session_thread_control::downstream_run(){
 	if( 0 < num_of_core_uses ){
 		CPU_ZERO( &mask );
 		for( int i = 0; i < num_of_core_uses; ++i ){
-			CPU_SET( static_cast<int>( i ), &mask );
+			CPU_SET( i, &mask );
 		}
 		sched_setaffinity( 0, sizeof( cpu_set_t ), &mask );
 	}
